Bring reducers up to date
