+++ src/reducer.spec.ts
@@ -1,0 +1,58 @@
+import { reducer, combineReducers, actionCreator } from 'rxbeach';
+import test from 'ava';
+import { marbles } from 'rxjs-marbles/ava';
+
+const throwErrorFn = (): number => {
+  throw 'error';
+};
+const incrementOne = actionCreator('[increment] one');
+const incrementMany = actionCreator<number>('[increment] many');
+const decrement = actionCreator('[increment] decrement');
+const handleOne = reducer(
+  incrementOne,
+  (accumulator: number) => accumulator + 1
+);
+const handleMany = reducer(
+  incrementMany,
+  (accumulator: number, increment) => accumulator + increment
+);
+const handleDecrement = reducer(decrement, throwErrorFn);
+
+const inputs = {
+  '1': incrementOne(),
+  '2': incrementMany(2),
+  d: decrement(),
+};
+const outputs = {
+  '2': 2,
+  '4': 4,
+  '5': 5,
+};
+
+test('reducer should store reducer function', t => {
+  t.deepEqual(handleDecrement, [decrement, throwErrorFn]);
+});
+
+test(
+  'combineReducers should reduce actions to state',
+  marbles(m => {
+    const source = m.hot('  121', inputs);
+    const expected = m.hot('245', outputs);
+
+    m.expect(
+      source.pipe(combineReducers(1, [handleOne, handleMany]))
+    ).toBeObservable(expected);
+  })
+);
+
+test(
+  'combineReducers should not silence errors',
+  marbles(m => {
+    const source = m.hot('  1d1', inputs);
+    const expected = m.hot('2# ', outputs);
+
+    m.expect(
+      source.pipe(combineReducers(1, [handleOne, handleDecrement]))
+    ).toBeObservable(expected);
+  })
+);
