+++ src/reducer.ts
@@ -1,48 +1,25 @@
 import { OperatorFunction, pipe } from 'rxjs';
 import { scan } from 'rxjs/operators';
 import { ActionCreator } from 'rxbeach';
-import { VoidPayload, AnyAction, UnknownAction } from 'rxbeach/internal';
-import { ofTypes } from 'rxbeach/operators';
 import {
-  ActionCreatorWithoutPayload,
-  ActionCreatorWithPayload,
-} from './types/ActionCreator';
+  VoidPayload,
+  UnknownAction,
+  UnknownActionCreator,
+} from 'rxbeach/internal';
+import { ofType } from 'rxbeach/operators';
 
 export type Reducer<State, Payload = VoidPayload> = (
   previousState: State,
   payload: Payload
 ) => State;
 
-type ReducerEntry<State, Payload> = [string, Reducer<State, Payload>];
+export type ReducerEntry<State, Payload = any> = [
+  UnknownActionCreator,
+  Reducer<State, Payload>
+];
 
-// Typescript is not so great with conditional types (ActionCreator) yet, so we
-// need to overload for each case
 /**
  * Create a reducer entry for passing into `combineReducers`
- *
- * This function doesn't alter the passed the reducer, but serves as a nice way
- * of typing it.
- *
- * @param actionCreator The action creator to assign this reducer to and extract
- *                      payload type from
- * @param reducer The reducer function
- * @template `State` - The state the reducer reduces to
- * @returns A reducer entry; A tuple array of action type and reducer, for
- *          passing into `combineReducers`
- *
- * @see combineReducers
- */
-export function reducer<State>(
-  actionCreator: ActionCreatorWithoutPayload,
-  reducer: Reducer<State, VoidPayload>
-): ReducerEntry<State, VoidPayload>;
-/**
- * Create a reducer entry for passing into `combineReducers`
- *
- * This function doesn't alter the passed the reducer, but serves as a nice way
- * of typing it.
- *
- * **NB!** If you need void payload, just skip the second type argument
  *
  * @param actionCreator The action creator to assign this reducer to and extract
  *                      payload type from
@@ -56,19 +33,12 @@ export function reducer<State>(
  *
  * @see combineReducers
  */
-export function reducer<State, Payload>(
-  actionCreator: ActionCreatorWithPayload<Payload>,
-  reducer: Reducer<State, Payload>
-): ReducerEntry<State, Payload>;
-/**
- * Untyped version. Your IDE should show you docs for the appropriate overload
- */
-export function reducer<State, Payload>(
+export const reducer = <State, Payload = VoidPayload>(
   actionCreator: ActionCreator<Payload>,
   reducer: Reducer<State, Payload>
-): ReducerEntry<State, Payload> {
-  return [actionCreator.type, reducer];
-}
+): ReducerEntry<State, Payload> => {
+  return [actionCreator, reducer];
+};
 
 /**
  * Combine reducer entries into a stream operator
@@ -85,17 +55,20 @@ export function reducer<State, Payload>(
 export const combineReducers = <State>(
   seed: State,
   reducers: ReducerEntry<State, any>[]
-): OperatorFunction<AnyAction, State> => {
-  const reducerMap = new Map(reducers);
+): OperatorFunction<UnknownAction, State> => {
+  const reducersByActionType = new Map(
+    reducers.map(([action, reducer]) => [action.type, reducer])
+  );
   return pipe(
-    ofTypes(...reducerMap.keys()),
-    scan((state: State, { type, payload }: UnknownAction) => {
-      const reducer = reducerMap.get(type);
-      if (reducer) {
-        return reducer(state, payload);
-      } else {
+    ofType(...reducers.map(([action]) => action)),
+    scan((state, { type, payload }: UnknownAction) => {
+      const reducer = reducersByActionType.get(type);
+      if (reducer === undefined) {
+        // This shouldn't be possible
         return state;
       }
+
+      return reducer(state, payload);
     }, seed)
   );
 };
